+++ chaoslib/action.py
@@ -4,7 +4,7 @@ import sys
 from chaoslib.activity import ensure_activity_is_valid, run_activity
 from chaoslib.exceptions import FailedActivity, FailedAction, InvalidActivity,\
     InvalidAction
-from chaoslib.types import Action
+from chaoslib.types import Action, Secrets
 
 __all__ = ["ensure_action_is_valid", "run_action"]
 
@@ -18,9 +18,9 @@ def ensure_action_is_valid(action: Action):
         raise InvalidAction(m).with_traceback(sys.exc_info()[2])
 
 
-def run_action(action: Action):
+def run_action(action: Action, secrets: Secrets):
     try:
-        return run_activity(action)
+        return run_activity(action, secrets)
     except FailedActivity as x:
         m = str(x)
         m = m.replace("activity", "action")