+++ chaoslib/probe.py
@@ -4,7 +4,7 @@ import sys
 from chaoslib.activity import ensure_activity_is_valid, run_activity
 from chaoslib.exceptions import FailedProbe, FailedActivity, InvalidActivity, \
     InvalidProbe
-from chaoslib.types import Probe
+from chaoslib.types import Probe, Secrets
 
 __all__ = ["ensure_probe_is_valid", "run_probe"]
 
@@ -18,9 +18,9 @@ def ensure_probe_is_valid(probe: Probe):
         raise InvalidProbe(m).with_traceback(sys.exc_info()[2])
 
 
-def run_probe(probe: Probe):
+def run_probe(probe: Probe, secrets: Secrets):
     try:
-        return run_activity(probe)
+        return run_activity(probe, secrets)
     except FailedActivity as x:
         if probe.get("negate") is True:
             return