+++ chaoslib/types.py
@@ -2,7 +2,7 @@
 from typing import Any, Dict, List, Tuple, Union
 
 __all__ = ["MicroservicesStatus", "Probe", "Action", "Experiment", "Layer",
-           "TargetLayers", "Activity", "Journal", "Run"]
+           "TargetLayers", "Activity", "Journal", "Run", "Secrets"]
 
 
 Action = Dict[str, Any]
@@ -17,3 +17,5 @@ TargetLayers = Dict[str, List[Dict[str, Any]]]
 MicroservicesStatus = Tuple[Dict[str, Any], Dict[str, Any]]
 Journal = Dict[str, Any]
 Run = Dict[str, Any]
+
+Secrets = Dict[str, Dict[str, str]]