+++ tests/test_probe.py
@@ -8,7 +8,7 @@ from chaoslib.exceptions import FailedProbe, InvalidProbe
 from chaoslib.probe import ensure_probe_is_valid, run_probe
 from chaoslib.types import Probe
 
-from fixtures import probes
+from fixtures import experiments, probes
 
 
 def test_empty_probe_is_invalid():
@@ -80,13 +80,13 @@ def test_http_probe_must_have_a_url():
 
 def test_run_python_probe_should_return_raw_value():
     # our probe checks a file exists
-    assert run_probe(probes.PythonModuleProbe) is True
+    assert run_probe(probes.PythonModuleProbe, experiments.Secrets) is True
 
 
 def test_run_process_probe_should_return_raw_value():
     v = "Python {v}\n".format(v=sys.version.split(" ")[0])
 
-    result = run_probe(probes.ProcProbe)
+    result = run_probe(probes.ProcProbe, experiments.Secrets)
     assert type(result) is bytes
     assert result.decode("utf-8") == v
 
@@ -96,7 +96,7 @@ def test_run_process_probe_can_timeout():
     probe["timeout"] = 0.0001
 
     with pytest.raises(FailedProbe) as exc:
-        run_probe(probes.ProcProbe).decode("utf-8")
+        run_probe(probes.ProcProbe, experiments.Secrets).decode("utf-8")
     assert "probe took too long to complete" in str(exc)
 
 
@@ -105,12 +105,12 @@ def test_run_http_probe_should_return_raw_value():
         m.post(
             'http://example.com', json=['well done'], 
             headers={"Content-Type": "application/json"})
-        result = run_probe(probes.HTTPProbe)
+        result = run_probe(probes.HTTPProbe, experiments.Secrets)
         assert result == ['well done']
 
         m.post(
             'http://example.com', text="['well done']")
-        result = run_probe(probes.HTTPProbe)
+        result = run_probe(probes.HTTPProbe, experiments.Secrets)
         assert result == "['well done']"
 
 
@@ -120,5 +120,5 @@ def test_run_http_probe_should_fail_when_return_code_is_above_400():
             'http://example.com', status_code=404, text="Not found!")
 
         with pytest.raises(FailedProbe) as exc:
-            run_probe(probes.HTTPProbe)
+            run_probe(probes.HTTPProbe, experiments.Secrets)
         assert "Not found!" in str(exc)