+++ CHANGELOG.md
@@ -2,7 +2,20 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/0.2.0...HEAD
+
+## [0.2.0][] - 2017-10-10
+
+[0.2.0]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/0.1.1...0.2.0
+
+### Added
+
+-   A first draft to loading secrets from environment
+-   Passing those secrets as a mapping to actions and probes expecting them
+
+### Removed
+
+-   Unused layer module
 
 ## [0.1.1][] - 2017-10-06